src/ai-service/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
openai-whisper==20231117
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
ollama==0.1.7
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1